Reinstantiate `-t` to be used as shorthand for `--trace`
According to the help text `-t` is used as short-hand to enable
`--trace`, while `-t` previously used for `--tiling` ended up being
renamed to `-T` in the `option` parser but not in the switch-case in
323a48b ("fix screenshot logic, and add tracing"), resulting in `-t` to
do nothing as the trace option was also using a zero-based long-option
value rather than the character literal `'t'`.  Fix that, as well as the
manual.